add first version with new test structure without start/stop volume
