+++ conf/example_usm.ini
@@ -12,5 +12,5 @@ usm_api_url = %(usm_web_url):9393/api/1.0/
 etcd_api_url = %(usm_web_url):2379/v2/
 usm_ca_cert = conf/tendrl_ca.crt
 usm_brick_name = brick
-usm_volume_name = TestVolume
 usm_pool_name = TestPool
+usm_id_fqdn = example-usm3-gl1.usmqe.tendrl.org
